// src/interfaces/apiResponseTypes/MedicalsReportsResponseType.ts
export default interface MedicalsReportsResponseType {
  heartDisease: boolean;
  hypertension: boolean;
  respiratoryDiseaseOrAsthma: boolean;
  diabetesMellitus: boolean;
  tuberculosis: boolean;
  epilepsyOrAnyNeurologicalDisorder: boolean;
  kidneyOrUrinaryDisorder: boolean;
  cancer: boolean;
  migraineOrPersistentHeadache: boolean;
  anyAllergies: boolean;
  disorderOfTheJointsOrMusclesArthritisGout: boolean;
  anyMajorSurgery: boolean;
  noneOfTheAbove: boolean;
}